+++ modules/siddhi-core/src/main/java/io/siddhi/core/exception/MappingFailedException.java
@@ -1,9 +1,30 @@
+/*
+ * Copyright (c) 2020, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
 package io.siddhi.core.exception;
 
-import io.siddhi.core.util.restream.model.ErroneousEvent;
+import io.siddhi.core.util.error.handler.model.ErroneousEvent;
 
 import java.util.List;
 
+/**
+ * Thrown during mapper level, when a payload is not able to be converted to a Siddhi event.
+ */
 public class MappingFailedException extends Exception {
     List<ErroneousEvent> failures;
