+++ modules/siddhi-core/src/main/java/io/siddhi/core/stream/StreamJunction.java
@@ -32,13 +32,13 @@ import io.siddhi.core.exception.SiddhiAppCreationException;
 import io.siddhi.core.stream.input.InputProcessor;
 import io.siddhi.core.stream.output.StreamCallback;
 import io.siddhi.core.util.SiddhiConstants;
+import io.siddhi.core.util.error.handler.model.ErroneousEvent;
+import io.siddhi.core.util.error.handler.util.ErrorOccurrence;
+import io.siddhi.core.util.error.handler.util.ErrorStoreHelper;
 import io.siddhi.core.util.event.handler.EventExchangeHolder;
 import io.siddhi.core.util.event.handler.EventExchangeHolderFactory;
 import io.siddhi.core.util.event.handler.StreamHandler;
 import io.siddhi.core.util.parser.helper.QueryParserHelper;
-import io.siddhi.core.util.restream.model.ErroneousEvent;
-import io.siddhi.core.util.restream.util.ErrorOccurrence;
-import io.siddhi.core.util.restream.util.ErrorStoreHelper;
 import io.siddhi.core.util.statistics.EventBufferHolder;
 import io.siddhi.core.util.statistics.ThroughputTracker;
 import io.siddhi.core.util.statistics.metrics.Level;
