+++ modules/siddhi-core/src/main/java/io/siddhi/core/stream/output/sink/Sink.java
@@ -28,10 +28,10 @@ import io.siddhi.core.util.ExceptionUtil;
 import io.siddhi.core.util.SiddhiConstants;
 import io.siddhi.core.util.StringUtil;
 import io.siddhi.core.util.config.ConfigReader;
+import io.siddhi.core.util.error.handler.model.ErroneousEvent;
+import io.siddhi.core.util.error.handler.util.ErrorOccurrence;
+import io.siddhi.core.util.error.handler.util.ErrorStoreHelper;
 import io.siddhi.core.util.parser.helper.QueryParserHelper;
-import io.siddhi.core.util.restream.model.ErroneousEvent;
-import io.siddhi.core.util.restream.util.ErrorOccurrence;
-import io.siddhi.core.util.restream.util.ErrorStoreHelper;
 import io.siddhi.core.util.snapshot.state.EmptyStateHolder;
 import io.siddhi.core.util.snapshot.state.State;
 import io.siddhi.core.util.snapshot.state.StateFactory;
@@ -435,6 +435,7 @@ public abstract class Sink<S extends State> implements SinkListener {
                     ErrorStoreHelper.storeErroneousEvent(siddhiAppContext.getSiddhiContext().getErrorStore(),
                             ErrorOccurrence.STORE_ON_SINK_ERROR, siddhiAppContext.getName(),
                             erroneousEvent, streamDefinition.getId());
+                    break;
                 case LOG:
                 default:
                     connectWithRetry();
