+++ modules/siddhi-core/src/main/java/io/siddhi/core/util/error/handler/model/AbstractErrorEntry.java
@@ -1,9 +1,30 @@
-package io.siddhi.core.util.restream.model;
+/*
+ * Copyright (c) 2020, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
 
-import io.siddhi.core.util.restream.util.ErroneousEventType;
-import io.siddhi.core.util.restream.util.ErrorOccurrence;
-import io.siddhi.core.util.restream.util.ErrorType;
+package io.siddhi.core.util.error.handler.model;
 
+import io.siddhi.core.util.error.handler.util.ErroneousEventType;
+import io.siddhi.core.util.error.handler.util.ErrorOccurrence;
+import io.siddhi.core.util.error.handler.util.ErrorType;
+
+/**
+ * Represents an entry which represents an error, in the {@link io.siddhi.core.util.error.handler.store.ErrorStore}.
+ */
 public abstract class AbstractErrorEntry {
     private int id;
     private long timestamp;
@@ -16,7 +37,9 @@ public abstract class AbstractErrorEntry {
     private ErroneousEventType eventType;
     private ErrorType errorType;
 
-    public AbstractErrorEntry(int id, long timestamp, String siddhiAppName, String streamName, String cause, String stackTrace, String originalPayload, ErrorOccurrence errorOccurrence, ErroneousEventType eventType, ErrorType errorType) {
+    public AbstractErrorEntry(int id, long timestamp, String siddhiAppName, String streamName, String cause,
+                              String stackTrace, String originalPayload, ErrorOccurrence errorOccurrence,
+                              ErroneousEventType eventType, ErrorType errorType) {
         this.id = id;
         this.timestamp = timestamp;
         this.siddhiAppName = siddhiAppName;
