--- modules/siddhi-core/src/main/java/io/siddhi/core/util/restream/util/ErroneousEventType.java
@@ -1,9 +1,0 @@
-package io.siddhi.core.util.restream.util;
-
-public enum ErroneousEventType {
-    COMPLEX_EVENT,
-    EVENT,
-    EVENT_ARRAY,
-    EVENT_LIST,
-    PAYLOAD_STRING
-}
