--- modules/siddhi-core/src/main/java/io/siddhi/core/util/restream/util/ErrorOccurrence.java
@@ -1,7 +1,0 @@
-package io.siddhi.core.util.restream.util;
-
-public enum ErrorOccurrence {
-    BEFORE_SOURCE_MAPPING,
-    STORE_ON_SINK_ERROR,
-    STORE_ON_STREAM_ERROR
-}
