--- modules/siddhi-core/src/main/java/io/siddhi/core/util/restream/util/ErrorType.java
@@ -1,6 +1,0 @@
-package io.siddhi.core.util.restream.util;
-
-public enum ErrorType {
-    MAPPING,
-    TRANSPORT
-}
