+++ modules/siddhi-core/src/test/java/io/siddhi/core/transport/TestSourceOptionSourceMapper.java
@@ -22,15 +22,17 @@ import io.siddhi.annotation.Example;
 import io.siddhi.annotation.Extension;
 import io.siddhi.core.config.SiddhiAppContext;
 import io.siddhi.core.event.Event;
-import io.siddhi.core.exception.SiddhiAppRuntimeException;
+import io.siddhi.core.exception.MappingFailedException;
 import io.siddhi.core.stream.input.source.AttributeMapping;
 import io.siddhi.core.stream.input.source.InputEventHandler;
 import io.siddhi.core.stream.input.source.SourceMapper;
 import io.siddhi.core.util.config.ConfigReader;
+import io.siddhi.core.util.error.handler.model.ErroneousEvent;
 import io.siddhi.core.util.transport.OptionHolder;
 import io.siddhi.query.api.definition.StreamDefinition;
 import org.apache.log4j.Logger;
 
+import java.util.Collections;
 import java.util.List;
 
 /**
@@ -61,7 +63,8 @@ public class TestSourceOptionSourceMapper extends SourceMapper {
     }
 
     @Override
-    protected void mapAndProcess(Object eventObject, InputEventHandler inputEventHandler) throws InterruptedException {
+    protected void mapAndProcess(Object eventObject, InputEventHandler inputEventHandler)
+            throws MappingFailedException, InterruptedException {
         if (eventObject != null) {
             if (eventObject instanceof Event) {
                 Event event = ((Event) eventObject);
@@ -69,8 +72,9 @@ public class TestSourceOptionSourceMapper extends SourceMapper {
                 event.getData()[1] = sourceType;
                 inputEventHandler.sendEvent((Event) eventObject);
             } else {
-                throw new SiddhiAppRuntimeException("Event object must be Event " +
-                        "but found " + eventObject.getClass().getCanonicalName());
+                throw new MappingFailedException(Collections.singletonList(
+                        new ErroneousEvent(eventObject, "Event object must be Event " +
+                                "but found " + eventObject.getClass().getCanonicalName())));
             }
         }
