+++ modules/siddhi-core/src/test/java/io/siddhi/core/transport/TestTrpSourceMapper.java
@@ -23,14 +23,15 @@ import io.siddhi.annotation.Extension;
 import io.siddhi.core.config.SiddhiAppContext;
 import io.siddhi.core.event.Event;
 import io.siddhi.core.exception.MappingFailedException;
-import io.siddhi.core.exception.SiddhiAppRuntimeException;
 import io.siddhi.core.stream.input.source.AttributeMapping;
 import io.siddhi.core.stream.input.source.InputEventHandler;
 import io.siddhi.core.stream.input.source.PassThroughSourceMapper;
 import io.siddhi.core.util.config.ConfigReader;
+import io.siddhi.core.util.error.handler.model.ErroneousEvent;
 import io.siddhi.core.util.transport.OptionHolder;
 import io.siddhi.query.api.definition.StreamDefinition;
 
+import java.util.Collections;
 import java.util.List;
 
 /**
@@ -83,10 +84,9 @@ public class TestTrpSourceMapper extends PassThroughSourceMapper {
                 }
                 inputEventHandler.sendEvent(output);
             } else {
-                throw new MappingFailedException("Event object must be either Event[], Event or Object[] " +
-                        "but found " + eventObject.getClass().getCanonicalName());
-//                throw new SiddhiAppRuntimeException("Event object must be either Event[], Event or Object[] " +
-//                        "but found " + eventObject.getClass().getCanonicalName()); // TODO check and remove
+                throw new MappingFailedException(Collections.singletonList(new ErroneousEvent(eventObject,
+                        "Event object must be either Event[], Event or Object[] " +
+                        "but found " + eventObject.getClass().getCanonicalName())));
             }
         }
     }
